pyproject.toml
[project]
name = "gpt_travel_planner"
version = "1"
dependencies = [
    "duffel-api",
    "ortools",
    "pyflightdata",
    "numpy",
    "pandas",
    "tenacity",
    "tqdm",
    "fire",
    "httpx",
    "haversine",
    "networkx",
    "python-dotenv",
    "functions-framework==3.*",
    "google-cloud-storage",
]

[project.optional-dependencies]
test = [
  "pytest",
  "black",
  "pytest_httpx"
]

dev = [
  "jupyter",
  "ipython",
]

[tool.black]
line-length = 88
skip-magic-trailing-comma = true

[tool.pytest.ini_options]
pythonpath = [
  "src"
]
markers = [
]

